Implement an API for sendMessage
